refactor: db 내부 객체 네이밍 수정
